Create Two_Fluid_relativistic.py
Created the code for a Two-fluid model with relativistic EoS for DM. In case someone wants to use a non-relativistic EoS for DM, there are instructions included inside the code as a comment on how to do that. You simply change the ex equation to the one of the comment.